[General] Fix bug with item usage

+++ src/ComboTracker.cs
@@ -22,7 +22,7 @@ namespace GCDTracker
             this.ComboUsed = new List<uint>();
             this.LastComboActionUsed = new(){0,0};
         }
-        public unsafe void onActionUse(byte ret, IntPtr actionManager, ActionType actionType, uint actionID, uint targetedActorID, uint param, uint useType, int pvp)
+        public unsafe void onActionUse(byte ret, IntPtr actionManager, ActionType actionType, uint actionID, long targetedActorID, uint param, uint useType, int pvp)
         {
             var comboDict = ComboStore.GetCombos();
