+++ src/Data/HelperMethods.cs
@@ -8,7 +8,7 @@ namespace GCDTracker.Data
 {
     static unsafe class HelperMethods
     {
-        public delegate byte UseActionDelegate(IntPtr actionManager, ActionType actionType, uint actionID, uint targetID, uint param, uint useType, int pvp, IntPtr a7);
+        public delegate byte UseActionDelegate(IntPtr actionManager, ActionType actionType, uint actionID, long targetID, uint param, uint useType, int pvp, IntPtr a7);
         public delegate void ReceiveActionEffectDetour(int sourceActorID, IntPtr sourceActor, IntPtr vectorPosition, IntPtr effectHeader, IntPtr effectArray, IntPtr effectTrail);
 
         public static bool IsWeaponSkill(ActionType actionType, uint actionID)
